Refactor DonationActivity and GoogleIAPHandler
